refactor(post_regression): remove t_obs and n_obs
This commit introduces a new dict attribute for the model,
called self.postreg. This dict stores user inputs from the
first call of the fit function: n_obs and t_obs, in a similar
way that fit_index is stored.

The functions model and post_regression are updated and the syntax
is considerably simplified, which should help code maintenance and
API usage.

The test_model.py is updated to consider that the fit function, from
the previous commits, use less parameters.

The notebooks SIR and SIR-X are updated to be consistent with these
changes.