feat: plot confidence intervals
The API plot_predictions was renamed to plot_pred_ci as
it is designed to provide short-term predictions and their
confidence intervals. The user should use predict for creating
long term predictions, although is adviced to not use the model
to predict out of the range of the length of the mean squared
error array produced by cross-validation.